Validate metadata of specialist documents correctly
This commit changes the validation of the specialist document
"metadata" hash.

The fix is quite interesting, because I think we've been
misunderstanding a part of the JSON Schema spec -- or at least I was.

Currently, the schema mandates that `details.metadata` in a specialist
document conform to one of ~15 subschemas, each conforming to one of
the specialist document types.

By using the `oneOf` keyword, we require that *exactly one* of the
subschemas validates. In practice, it's perfectly possible for the
metadata of an AAIB Report to be also valid for the metadata of a RAIB
report.

The`anyOf` keyword actually does what we want: it will make sure that
the metadata is valid against at least one schema.

The TIL is: you almost never want `oneOf`, unless you're doing
something to enforce mutually exclusive attributes.

More on the JSON Schema spec:

http://json-schema.org/latest/json-schema-validation.html#anchor88

This was discovered while trying to generate random content in the
`govuk_schemas` gem (https://github.com/alphagov/govuk_schemas_gem).