Don't allow `policies` to be set directly
Working groups can have a `policies` link type, which is populated with
the policies that have this working group in the `working_groups`
links. This is done by the dependency resolution functionality in the
publishing-api.

This makes clear that publishing apps should not use the link type, but
that it still can appear on the frontend.