Add parent to service manual topic example
Service manual topics will usually have the service manual homepage as a parent, so include it in the example.

The exception to this is when the topic should not be included on the list of topics on the homepage. A current use case for this is the ‘communities’ topic, which is linked separately.