Add generic schema
This commit adds a “generic” schema which includes nothing in the `details` or `links` hashes beyond what is defined in the base JSON documents.

The reason is to provide a schema for content items that do not need any information in the content-store to be rendered.

Previously, the `special_route` schema has been used for these types of pages, for example the “Find your local council” page (https://github.com/alphagov/frontend/blob/e460d40b0095d6c07570aa14b884fcbdde65b7c6/lib/tasks/publishing_api.rake). This is wrong, because `special_route` is a special kind of schema that doesn’t have a frontend representation. It was intended for the use of routing, never to be served from the content-store (this was never implemented). We think `special_route` could be deprecated in favour of this schema.

Other names we considered:

- base
- basic
- page
- minimal
- minimum