Reformat examples as part of the build
This adds a step to the build process to reformat the example JSON
files. The main reason is that this allows us to quickly edit all the
examples. For example,
https://github.com/alphagov/govuk-content-schemas/pull/444 removes the
`format` field simply by adding `data.delete("format")` to the rake
task.

This will also prevent errors with duplicated keys in the JSON
(converting it to a Ruby hash will solve them).