Add rake task to remove all V1 schemas again
At the moment the schema generation is quite complicated - the V2
schemas depend on the V1 schemas. This makes removing the code that
generates the V1 schemas pretty hard.

I've been working on a rewrite of the schema generation code. To make
that PR easier to review, this commit adds a rake task to delete all V1
schemas as a last step. It's quite silly.