Tweak example text for the CMA finder content item
When a new email alert subscription is created, a short name is generated describing the subscription. If enough subscription items are chosen by the user, the short name becomes longer than 255 characters, which is the GovDelivery limit for this field.

This commit tweaks the text for the finder content item examples to reflect changes made to the CMA case finder to keep the short text under 255 characters.