Add more text options for email alerts
This commit adds the ability to define a separate email alert list title text which will be used when the normally generated string is too long (> 255 characters). It will be appended with a string of the form “x items”, where “x” is the number of selected topics and “items” is either the singular or plural string defined as `email_filter_name`.