Migrate all CI build steps into Jenkinsfile
This removes the Jenkins 2 build's dependency on the old jenkins.sh.

The jenkins.sh file cannot be deleted yet because it is still used in
the Jenkins 1 build before the schema tests are triggered. It will be
safe to delete once all the downstream schema test builds have been
migrated to Jenkins 2.