Remove reason for change from change history definition
The "reason_for_change" field inside a change history definition was
added in #280 to support `service-manual-publisher`, which wanted both a
reason for change and a change note.

We're now removing the reason for change field as it's no longer
considered useful. This removes it from the shared definition.

In the 10 months that have elapsed since it was added, it doesn't appear
to be used by any other applications. Searching GitHub for non-JSON
references to "reason_for_change" outside of the
service-manual-publisher repo yields nothing:
https://github.com/search?utf8=%E2%9C%93&q=org%3Aalphagov+reason_for_change+-repo%3Aalphagov%2Fservice-manual-publisher+-language%3Ajson&type=Code&ref=searchresults