Rename fact-check-id
We want to be able to give people access to the draft stack for other states
than just fact check (i.e. draft or ready) so we rename the field to be less
specific. The [docs of authenticating
proxy](https://github.com/alphagov/authenticating-proxy#technical-documentation)
refer to this feature as "The application also supports bypassing
authentication via a valid JWT token" so the new name `auth_bypass_id` feels
appropriate here...

We are keeping the fact-check-id field as a deprecated field, until publisher
documents have been republished to use the auth_bypass_id field.