Lock down the allowed document types
This locks down the allowed values for `document_type` to a predefined
list. The list of produced from the live content-store, so this should
be backwards compatible.

Reasons for doing this:

- It becomes harder to inadvertently change the `document_type` of a
document. Most recently we accidentally changed the document type of a
number of mainstream documents to `generic_with_external_links`
(https://github.com/alphagov/publisher/pull/581). Because the document
type is used by analytics, this affected the GOV.UK performance
reporting.
- We can use the centrally defined list in the developer docs. We
currently rely on search for this, but that isn't accurate because
content items with a type like "redirect" or "gone" will never be in
the search index.
- The "document supertype" gem
(https://github.com/alphagov/govuk_document_types) can use this list to
validate the provided document types. This makes it easier to spot
errors.

Note that because the current list is taken from the live content store
it contains document types that are obviously wrong, such as the
`placeholder` and `placeholder_ministerial_role` documents. This has to
be fixed in the content store before removing them here.