Include featured policies in placeholder edition links
Organisations are a placeholder format on Whitehall. To render an
Organisation on Whitehall there is currently use of the Publishing API
linkable endpoint, this has proven nicht gut causing some errors in
production.

This allows placeholder documents to have edition links to
featured_policies so we can start working around this problem.

It seemed a shame that all placeholder items may have to deal with this,
but this already seems to be the case with link set links, so doesn't
appear to be a new precedent.