Add edition_links.json for DocumentCollection
Allow `documents` edition links for the DocumentCollection format. There
is currently some duplication in links that are in `base_links`
and `edition_links` but as other formats are dependent on the
`base_links` we can't currently remove them. It may be that they can be
moved to `base_edition_links` which will be done separately.