Updates schema for Html Publications
For: https://trello.com/c/88zaXS8J/47-create-print-styles-for-official-documents

This will add two more fields to be sent along to Publishing API from Whitehall. These two fields will be included in the print version of HTML Publications which are edited on whitehall, but will be displayed by frontend.

You can still see a draft of the HTML Publication on Whitehall, but the frontend app actually needs to display this information.

These schemas will signal to frontend to use two extra fields (that will be displayed in the print version of the HTML Publication).

They are: isbn and print_meta_data_contact_address. Both fields are optional.

isbn - Identifies the Web ISBN to be displayed when printing an HTML Publication
print_meta_data_contact_address - Identifies the address of the institution which has produced the HTML Publication