Merge lists of allowed document types
Remove the comments about how the allowed document types were generated
because govuk-content-schemas is now the source of truth about document
types rather than the content store.

Combine this lists of allowed types for the same reason: this list
defines the allowed types, so there's no distinction between types which
were in the content store when the list was generated and new types.