Extract a definition for will_continue_on
Both licences and transactions have this field, so tests fail until we create a
definition for them. We considered creating a definition for an object that
would encompass both the will_continue_on and the external_link it is referring
to, however the rule of "one field in the admin in the publisher = one
attribute of the schema" prevailed.