Send current commit sha to downstream jobs.
Set a flag on those apps that are using the centralised buildProject
script. If so, don't wait for them to complete; they will report their
status individually to the triggering PR.

Removed the reporting of times to statsd as it will not work with
jobs that do not wait for completion.