Build frontend schemas using v2 schemas
When we added links into v2 publisher schemas we hit a point where the
v1 schemas were no longer suitable to generate the frontend schemas.

This changes the approach to frontend schema generation to make use of
the publisher_v2 schema.json and links.json as means to generate the
frontend schemas.

We only use one source to generate the frontend schemas as the different
source files are not interchangeable.