Rewrite schema generation
This rewrites the schema generation to be easier to follow, faster and
easier to extend.

In this PR we go out of our way to keep the /dist output in of the
rewrite exactly the same as the old generator. This is to make sure
that we aren't changing the functionality inadvertently and to make
reviewing easy. In a follow-up PRs we can make some of the code nicer
and refactor some more.