Add more allowed document types to `special_route`
Add `homepage` and `search` as allowed values for `document_type`. This
is a temporary measure to allow us to re-present the homepage and search
content items with custom document types so that they pick up the
correct supertype for analytics.

A later PR will fix this properly by replacing the `special_route`
hard-coded schema with one which is auto-generated from components like
most of the other schemas.