Move special route example into example directory
Move robots.txt.json from publisher_v2 to publisher_v2/examples because
it is an example of a special route rather than part of the
`special_routes` schema.

This file used to be in an examples directory, but it was moved into a
schema directory by mistake when the v2 schemas were first generated.