Add locale and details to hard-coded special_route schema
Add optional properties to the `special_route` publisher schema which
match the properties which will be generated when `special_route` is
migrated from a hard-coded schema to an automatically generated one.

This is step 1 of a process of making `special_route` consistent with
the other schemas, which will allow it to use the globally-defined list
of allowed document types:

1. Add `locale` and `details` as optional properties, and hard-code
their definitions
2. Update other applications to provide these properties when publishing
a `special_route`
3. Replace the hard-coded `schema.json` with the schema components like
`metadata.json`, which will make `locale` and `details` mandatory
fields.