Refer to downstream builds using absolute rather than relative path
This fixes the triggering of the info-frontend build, which was failing
with the message `ERROR: No parameterized job named
info-frontend/deployed-to-production found`.

The problem was that the branch `info-frontend` has the same name as the
`info-frontend` Jenkins project. When Jenkins tries to run `build job:
"info-frontend/deployed-to-production"`, it looks in the local directory
first and tries to build a (non-existent) job under the
govuk-content-schemas `info-frontend` branch rather than building the
info-frontend project.