Add world_location schema
WorldLocations are no longer going to be rendered by Whitehall (with the
exception of the international delegation type) as they are going to be replaced
by taxons. They do still need to be linkable by other content though as
other things depend on this assocation e.g. email alerts.

They are currently sent as placeholders to the publishing API. This
commit creates a new schema for them (which is just a placeholder copy
with the name changed) and then removes the requirement for a base path
from  that schema.