Add base_path_less_frontend_links definition
Some formats can be base_path less. These formats will have a minimal
expanded link representation.

This commit adds a definition for that and applies it currently to just
`world_locations` links. `contacts` and `governments` are also base_path
less but they aren't represented in the examples so it is unclear where
these occur. They can probabaly also use this definition but I haven't
added them as part of this work.

As some base_path-less links can still have a base_path, web_url etc
these have been included but are not required in the definition.