Apply a consistent ordering to schema fields
This makes schema outputs consistent and is especially useful for
tinkering with how schemas are generated since it's far easier to
compare the outputs of consistently defined schemas.

There is a somewhat complicated set of rules to define the ordering
which are there to optimise schema readability. Using an alphabetical
approach would be the most straight forward but you end up with
"definitions" swamping schemas and "type" getting lost.