Add special_route example
This commit adds an example of the special_route format. This format
isn't rendered but rendering apps do need to handle it now that
content-store will fall through to the first prefix route that it finds
if a specific content item is not found.

e.g. `/government/i-dont-exist` will return the `/government`
special_route item.

Having this example will allow us to test for this and handle it
properly.