Make payload version mandatory for notifications
When we added the payload_version field to the notification schemas, we made it
optional because it needed to be compatible with the old version of the
publishing API (which did not add the field) and the new version (which does).

Rummager is now relying on the presence of that field and the publishing API
always adds it, so we should make it a required field.

https://trello.com/c/i2ZJtwyQ/220-make-payloadversion-a-required-field