Re-introduce world_locations frontend links hack
This re-introduces a messy hack that meant world_locations could be
linked to without a base_path. The reality is any link type could
potentially not have a base_path, and link_type tells us nothing about
the document.

This is re-introduced to reduce the barriers to get this PR merged and
hopefully can be removed soon.