Add whitehall-admin as a rendering app
whitehall-admin, rather than whitehall-frontend is responsible for
responding to requests for /government/uploads. Currently,
govuk-puppet sets up the NGinx configuration such that when the
request ends up at the frontend load balancer, it reverse proxies the
request to whitehall-admin.

Adding whitehall-admin as a valid rendering app will allow creating a
special route for directly routing traffic to whitehall-admin, which
will help with simplifying the routing, and making it more explicit.