Break up the monstrous definitions/all.jsonnet file
This file was a bit of a nightmare to browse due to it's length, and
it's presence was likely to encourage using it as some sort of dumping
ground for definitions.

This groups definitions based on their contents - I hope it's usable but
I could imagine it's difficult to work out which file would have the
definition you want.