Handle document_type with a whitelist
This restores `lib/govuk_content_schemas/allowed_document_types.yml`
which is used by other projects and uses that as a means of providing a
whitelist of all the allowed document type values.

Rather than having a definition file for document_types we now build
this definition automatically based on the contents of this YAML file,
we also verify that values given in schemas are within this list.

Finally we wrap errors with formats in a InvalidFormat class which can
be caught and inform the user which schema file has the problem.