Create new file for document types
This YAML file also has humanised names (the names capitalised, and
abbreviations corrected). We'll use this name in the docs first. We can
also use it to better communicate the intent of the document type.

Making the YAML file a hash allows us to add more metadata like
descriptions and status later.