Integrate dummy content store
This commit integrates the dummy content store
(https://github.com/alphagov/govuk-dummy_content_store).

Differences between this and the gem:

- It's no longer a gem so we don't have to release a new version for
each change and bump the version. This makes it easier for the app on
Heroku to keep up to date (the app is deployed from this repo to make
sure that the schemas used by the app are always up to date).
- By using Sinatra we can drop a lot of the code needed to render
responses and render the ERB template. This version isd hundreds of
lines smaller.
- Instead of proxying the content store, this app redirects to the live
content store. This removes complexity and is more robust.
- Makes sure that items returned have the correct base_path.
Applications like government-frontend sometimes look at the base_path
to determine if they should redirect to a more canonical URL.

Tested by running `./startup.sh` in this repo and then this for
government-frontend:

```
PLEK_SERVICE_CONTENT_STORE_URI=http://localhost:4567/api ./startup.sh
--live
```

After this is merged govuk-dummy_content_store can be archived.