Fix generated schema for calendar
https://github.com/alphagov/govuk-content-schemas/pull/677 and and
https://github.com/alphagov/govuk-content-schemas/pull/678 were merged
around the same time, so the calendars schema didn't have all the
changes. This is causing subsequent PRs like
https://github.com/alphagov/govuk-content-schemas/pull/680 to fail.

This situation can be prevented by requiring branches to be up to date
before merging (there's a GitHub setting for that), but it's so rare
that I don't think it's worth the hassle.