Add external_content document type
This document type will represent related external URLs such as local
council websites.

This content was previously just stored in the search index, but this
change will allow us to store it in the publishing API so that the
publishing API can send the content to rummager rather than a publishing
app.