Add external_content schema
Add schema which represents links to content which is relevant to GOV.UK
but is not hosted there, such as local councils or selected NHS pages.

These URLs are indexed in site search so that we can direct users to
relevant non-GOV.UK content.

These URLs are currently sent to rummager (and the search index) by the
search admin tool. This new schema will allow search admin to send pages
to the publishing API instead, which now sends all publishing events to
rummager. This will help us reduce the number of direct dependencies on
rummager.