Add residential_property_tribunal_decision document type
This is a new specialist document for MOJ / HMCTS with a category,
sub-category and decision date as allowed fields.  Unlike other tribunal
decisions we add the explicit allowed values for the sub category to
the schema (it's unclear why the other tribunals haven't).  The values for
the sub-category include the category they belong to as a prefix, and are
separated from the sub-category by a triple-dash.  This is to make it
easier to validate that the category and sub-category match up in
specialist publisher.