Add JSON Schema keywords documentation
The Modelling Services team have been working on the content schema for `step by step navigation` (previously called `task lists`).

We had to do a bit of research to find out how to represent certain properties of our task list, and thought it would be good to add to our documentation.

Please feel free to add to this.