Include redirects in frontend redirect schemas
Currently it's possible to have a redirect make it's way to a frontend
application (for example when a route is in the process of changing to a
redirect at the time of a request). Frontend apps don't have any means
to deal with these redirects as they are not provided with the content
item.

Since this field is only useful on redirects it's only enabled on
documents with that schema name. It is also an optional field as any
existing published redirects won't have it.