Add fields to step by step navigation
We need to be able to display a title and introduction to a step by step
navigation.  These may be different to the page title and the description
(which is used as the meta description and in search).  The introduction may
also be in markdown as we've already identified that this will need to have
several paragraphs for future step by step navigation journeys.