Add new organisation schema
This commit adds a new organisastion schema that will be used when populating organisations in the content-store. This will allow everything that is currently represented on these pages to be stored in the content-store. It also deprecates a link type in the placeholder schema that will be replaced by an equivalent in the organisation schema.