Make certain organisation fields nullable
This commit allows `name_prefix` and `role_href` for people and `public_updated_at` and `document_type` for featured documents be `null`, since these fields are not always expected to be populated. It also moves the person definition into a shared file since all person definitions for organisations are identical.