Add role appointments
This commit adds role appointments as a new schema. This schema is basepathless, and will model the relationship between people and roles, including both current and previous roles.

It also removes the `role_type` field from the role schema, which is unused and can be determined by the `document_type`.