Expand Whitehall summary_organisations
We need more information in `summary_organisations`.
The reason behind this is that we need to maintain the `/api/organisations` current endpoint, this endpoint is managed by
Whitehall::Api and since organisations are being extracted from Whitehall we are moving this functionality to be rendered by collections app.

The main objective is to keep the endpoint exactly as it is but being fetched from content-store.
Without this change, the alternative is to find each content_item (organisation) by base_path and retrieve the necessary information,
although this would cause several more requests. We think this is the best approach.

Trello: https://trello.com/c/p7ywlbk1/130-migrate-organisations-api