Speeches can be tagged to policies
Although "related_policies" is also a thing, and seems to duplicate "policies",
there are items in the content store that use "policies". I don't know what
link types should and shouldn't be used, but for now I'm expanding the schemas
to describe reality rather trying to fix reality.

Example content item that is not currently valid:
https://www.gov.uk/api/content/government/speeches/caroline-spelman-addresses-the-abi-conference-on-climate-change-and-flood-risk