Revert "Speeches can be tagged to policies"
This reverts commit af806989644fa31591532fddb857b5d646c21073.

Policies should not appear in any publishing schema according to
https://github.com/alphagov/publishing-api/blob/a053615339cf4637d44a574adf426639531b1de0/spec/lib/expansion_rules_spec.rb#L209

This constraint is to avoid "magic" links overriding manually set ones.

"policies" is normally a reverse link of "working_groups", and "speech" is the
only kind of document where we've manually tagged to it.

Since there is also a "related_policies" link type the way forward is
probably to delete the "policies" links rather than fix the schema.