Duplicate news article links as edition links
The new Content Publisher app will take the approach of using edition
links wherever possible, the key advantage being that edition links are
part of the published content and can therefore be scheduled/versioned.

Initially we would like to support a subset of the current links, such
as world_locations and organisations; other kinds of links, such as
roles and people, are technically more challenging and will come later.

Note this should have no effect on current publishing tools, as existing
linkset links will be used if edition links do not exist for a document.