Rename Service Standard
The Digital Service Standard is being renamed to the Government Service Standard. The number of points in the standard will also be reduced from 18 to 14.

Whilst these are only examples, changing these allows us to be consistent in the tests in service manual frontend and service manual publisher, which rely on these example schemas.