Add roles and people to news edition links
This makes these two link types available as edition links so that they
can be used by Content Publisher.

It also updates the descriptions to not be email specific since these
aren't only used for email.

These links do seem a bit fishy as they all used to link to both aspects
of a role appointment without the link to the role appointment. It would
probably make more sense that these linked to a role appointment and
these were expanded. Though this would be a more substantial change.