Allow bulk_publishing parameter for Publishing API payloads
Despite this field having been available in Publishing API for 2 years
it's not actually accepted at a schema level. Presumably, in the case of
PutContent, this means it's not been used in the last couple of years as
it would have returned a 422 response.

As it's quite a useful option to have I felt it was better to add it
into the schemas rather than remove the feature from Publishing API.