Remove requirement for base_path in linked items
Org pages link to several things that don't have base_paths, such as
organisation contacts and foi contacts.

In theory there is no restriction on which content types are linked, so this
check feels a bit like a false safety net because a content item without a base
path could be added to any link section.  In practice, this doesn't seem to
happen.

I tried removing the requirement for a base path from this section, but a lot
of downstream tests failed on apps that use govuk_schemas to generate example
content items.  I don't have time to fix them all at the moment so I'll reuse
the existing pattern here instead.