Add iso2_country_code field to contact post addresses
This is used to determine how an address is rendered on the frontend as
different countries have different styles.

This is added in addition to world_location field since it's not
particularly reliable to try determine this from that. However if we'd
started with this in the first place we could potentially have dropped
that field and derived the value using this.