Stop requiring government/political for news
Currently each news article must specify whether t is political and
irrespective have an associated government. The logic to do this in
Whitehall is complex and was implemented in a rush when there was a
change of government; we don't want to do this in Content Publisher.

Rather than send junk or inaccurate data to the Publishing API, it seems
we can avoid specifying these fields without causing issues downstream.

  * https://github.com/alphagov/finder-frontend/blob/master/app/views/finders/_results.mustache
  * https://github.com/alphagov/government-frontend/blob/7942c6f20dfac7f99c572b9a20e3c4fb7463c8c4/app/presenters/content_item/political.rb
  * https://github.com/alphagov/govuk_publishing_components/blob/master/lib/govuk_publishing_components/presenters/meta_tags.rb

Indeed, with the current implementation in Whitehall, we need to
manually republish all political content when the government changes.