Add variants for transactions
This commit adds a `variant` field to the `transaction` schema. Each `variant` has a number of fields related to those of the parent content item, as well as a `slug`. When a variant of a transaction is accessed via `/name-of-transaction/slug-of-variant`, the fields that are defined for the variant are rendered in preference to those of the parent content item.