Add taxonomy filtering
This allows us to set one or more taxonomy topics as a base filter and use them
in an AND style query with another taxonomy topic.

We only need it for one topic at the moment, but it'd be tricky to expand later
so we'll start with allowing more than one.