Add a "combine_mode" to finder facets
Say you have a business in the agriculture sector, and you export to
the EU.  If "sector" and "exporting" are facets, then the current AND
combining works:

    sector && exporting

If you also employ EU citizens, and that's a new facet, then you get
this situation:

    sector && exporting && employing

Which is unlikely to return what you want.  Why would a document be
tagged to all three?

What you actually want is something like this:

    sector && (exporting || employing)

This PR adds a field to each facet to say how to combine it.  The
field is optional.