Format the example files
There are a bunch of examples with funky indentation and weirdly
formatted keys.

This makes formatting the examples part of the build process. It'll
make the examples more consistent, but also makes it possible to
mass-edit the examples by putting some code in the rake task. We'll use
this to get rid of the supertypes in the frontend examples
(https://github.com/alphagov/govuk-content-schemas/pull/852).