Remove supertypes from frontend schema
This commit removes the supertypes from the allowed output of the
content store. We're removing it from the content items in favour of a
system where we get the supertype of a document from directly from the
gem: https://github.com/alphagov/govuk_document_types.

This will make it easier to iterate the supertypes. We won't have to
republish everything if a new type is created or one is modified.

We are keeping the notification schemas the same for now, because the
content performance manager project still uses the values. Once those
are removed, we'll retire the groups here entirely.