Add workflow_message to notification properties
This is a message passed in the payload to the notify endpoint in the
publishing API, it overrides the change note used in the notification
in cases where we explicitly need to explain why the notification was
sent and what the content or workflow change was, eg. tagging to a finder.