Wait for dependant application tests to be run
And propagate the build result, so that the build results for
dependant applications are taken in to account for the GitHub branch
protection configuration.

Previously, as the main job is the only one taken in to account for
the branch protections, Pull Requests could be merged when the changes
were incompatible with applications.