Add current field
This will be set to true if the role appointment is "current". Another
way of finding this out is to look at whether then `ended_on` field is
set, but I thought it would be good to pull this out as denormalised
data to simplify the logic of the frontends.