Add person_appointment_order field
This is a new field on role appointments which can be used to order
appointments with the same `started_on` date by "importance".

This is used to order the appointments in some places on the frontend,
notably the title which gets shown above the name of the person.

An alternative to putting this into the role appointment would be a
field in the person, but that adds a dependency between the person and
the role appointment meaning that if a new role appointment is created,
a new edition of a person would need to be drafted to include the new
role appointment in the details.