Allow links to governments in some schemas
This is meant to replace the government property in the details. Using
the link expansion functionality in the Publishing API will provide a
more effective way to manage the relationship between content and a
government.

Linking all content in bulk is easiest with normal links, but I've
included government within the edition links to keep that option open
as well.