Add optional "attachments" list field to Whitehall formats
This list will become mandatory when Whitehall and Content Publisher
are publishing docs with the field set.

These are all the formats which Whitehall allows attachments for:

    $ grep Attachable app/models/* | grep include | cut -d: -f1 | uniq
    corporate_information_page.rb
    detailed_guide.rb
    news_article.rb
    policy_group.rb
    publicationesque.rb
    response.rb
    world_location_news_article.rb

policy_group.rb corresponds to the working_group schema.

response.rb doesn't publish to the publishing api, so it has no
schema.

publicationesque.rb defines a class which is referenced in:

    $ grep Publicationesque app/models/* | cut -d: -f1 | uniq
    consultation.rb
    publication.rb
    publicationesque.rb
    statistical_data_set.rb
    world_location.rb