Add global fields to homepage schema
This adds a global field to the homepage schema. The purpose of this
is to hold information that is used to render the consistent header
and footer across GOV.UK. The usage intention is that the global data
held in the homepage can be shared with every single GOV.UK content item
so that they no longer need to rely on static as a source for shared
global configuration.

This work has been done as part of a spike to explore the possibility of
removing static and is a prototype of a concept for modelling global
data. I chose to break this data up into header and footer sections. In
the header the only consistent global data seemed to be the emergency
banner and global bar (which I renamed to alert bar). In the footer
there are a lot of links in groups.

I'm not too sure how much consistency and flexibility we want in terms
of links in the footer. My instinct was to expect a footer component to
define which groups of links are acceptable and then have the actual
links configured by the homepage data. This seemed to offer an easy
ability to change a link or even remove a whole section (by providing an
empty array) but avoided the more complex concerns of having to consider
rules about how groups of links are presented.

As the templates for this have a lot of hardcoded links I decided it was
best to set up some basic schemas for defining links which could be used
outside of this situation and could be expanded for any extra
attributes. These only support links where we know the href. A potential
future enhancement would be to devise a way that these links can use
data from a content items expanded links, however this is a much
trickier problem.

Things that'd be good to get input on for this:

- are there better names than global?
- is there any other data needed to render a GOV.UK page that this is
  missing?
- should we try have links in a content item or is this better as
  something in a component (that'd need co-ordinated deploys)
- should we provide more customisation for the links? I notice in the
  time since I started this a new footer link has been created for the
  coronavirus